third_party/blink/web_tests/external/wpt/html/cross-origin-embedder-policy/credentialless/dedicated-worker.tentative.https-expected.txt

This is a testharness.js-based test.
PASS dedicated-worker
PASS fetch same-origin + credentialless worker
FAIL fetch same-origin assert_equals: coep:credentialless =>  expected "Worker blocked" but got "same_origin"
FAIL fetch cross-origin assert_equals: coep:credentialless =>  expected (string) "Worker blocked" but got (undefined) undefined
FAIL fetch cross-origin + credentialless worker assert_equals: coep:none =>  expected (undefined) undefined but got (string) "cross_origin"
Harness: the test ran to completion.
